Extract Episode_Block_Tags helper
